+++ CHANGELOG.rst
@@ -6,6 +6,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on `Keep a Changelog <https://keepachangelog.com>`_.
 
+0.4.0
+------
+
+Added
++++++
+- `infectivity_or_neutralized` option to allow fitting of fraction neutralized as well as fraction infectivity.
+
+Fixed
++++++
+- `scipy` deprecation warnings.
+
 0.3.1
 ------